fix: update EOA connections to re-call connect() funciton on re-try
